fix(ken-414): say what the x11 pin is really for
The branch has attributed the AppImage's GDK_BACKEND=x11 pin to the WebKit
DMABUF crash since its first commit, and that is wrong. tauri#8541, the
issue the bundled hook's own comment cites, is a GLib-GIO settings-schema
lookup that aborts the process on bundles built old and run new; I read the
issue rather than take the claim, and it also records that the abort does
not occur for bundles built on current Ubuntu. WEBKIT_DISABLE_DMABUF_RENDERER
cannot touch a schema lookup, and the two failures do not substitute for
each other. The rustdoc, ARCHITECTURE and the CHANGELOG all said otherwise
and now say this. The pushed commit bodies carry the old claim and are left
alone: rewriting the branch under an open review costs more than the record
of a correction that this commit is.

The wayland,x11 list is described as what it is: a fallback for a display
that cannot be opened, not a net under an abort. A g_error kills the
process after GDK has chosen Wayland, so the second entry never runs.

I did not build the supervising spawn. Turning the exec into a parent that
watches a child and retries on x11 would trade a mechanism that cannot fail
for a heuristic about whether a window ever appeared, shipped untested
because neither of us can reproduce the abort, and it would double the
process tree the launcher and window manager see - all for a failure that
upstream's own report says does not reach bundles built the way release.yml
builds ours, and that the released AppImage patched to wayland,x11 did not
show on this host. The recourse is KENDEX_GDK_BACKEND=x11, and it is now
documented in all three places, including the CHANGELOG so a person hitting
it has somewhere to look. If a report ever arrives, that is the point to
build the supervisor, with a reproduction to test it against.

The reveal in setup has no test and does not get a fake one. Tauri's mock
runtime answers is_visible with a constant and show() does nothing, so a
mock test would pass whether or not the window is ever shown. What does
catch the deletion is clippy: zoom has no reader but that closure, so
removing it fails -D warnings on an unused variable. Verified by deleting
the line. That is incidental rather than designed, so the line now says so,
and says what would take the protection away.

A late older resize that succeeds after a newer one was refused left the
store behind the window: the refusal rolled back past a size the window
then landed on, and the commit wrote the rolled-back one. The last reply
standing now reconciles the store, with a test that drives the replies in
that order.

The zoom save suite crossed the line cap and splits along the seam it
already had: what the commit writes while a resize is still out, and the
write mechanism itself.
